refactor: remove tool and context gateways
