+++ docs/artifacts.md
@@ -26,106 +26,6 @@ Current storage layout:
   - Session/manifests DB: `~/.rakh-dev/sessions/sessions.db`
   - Blob storage: `~/.rakh-dev/artifacts/blobs/sha256`
 
-## Temporary tool artifacts
-
-ToolGateway also uses a separate temporary artifact store for oversized tool
-results. These are not durable session artifacts and they are not shown in the
-normal artifact pane.
-
-Current storage layout:
-
-- Release builds:
-  - Tool-artifact root: `~/.rakh/artifacts/tools`
-- Debug/dev builds:
-  - Tool-artifact root: `~/.rakh-dev/artifacts/tools`
-
-Each temporary tool artifact lives in its own directory named
-`<timestamp>_<random>`. The directory contains:
-
-- `meta.json`
-- `payload.txt`
-
-The metadata currently records:
-
-- `artifactId`
-- `createdAtMs`
-- `runId`
-- `tabId`
-- `agentId`
-- `toolCallId`
-- `toolName`
-- `sourceKind`
-- `policyId`
-- `originalFormat`
-- `sizeBytes`
-- `lineCount`
-- optional `intention`
-
-Temporary tool artifacts are always canonical UTF-8 text:
-
-- plain string tool results are stored directly
-- structured results are pretty-printed as JSON text
-
-The agent never receives raw filesystem paths. It only sees logical
-`artifactId`s through ToolGateway envelopes.
-
-### Tool-artifact helpers
-
-The model-visible helpers for temporary tool artifacts are:
-
-- `agent_tool_artifact_get`
-- `agent_tool_artifact_search`
-
-The backend also exposes create/delete operations, but those are engine-owned
-helpers rather than normal model-facing tools.
-
-`agent_tool_artifact_get` returns a bounded slice of the stored payload and can
-optionally read by line range.
-
-`agent_tool_artifact_search` performs regex search over the stored payload and
-returns bounded contextual matches.
-
-These tools intentionally enforce conservative caps so reading a large artifact
-does not immediately recreate the same context-size problem.
-
-### ToolGateway envelopes
-
-When ToolGateway replaces a large tool result, the model receives an artifact
-reference envelope instead of the original payload:
-
-```json
-{
-  "ok": true,
-  "data": {
-    "__rakhToolGateway": {
-      "kind": "artifact-ref",
-      "artifactId": "toolart_deadbeef",
-      "originalTool": "exec_run",
-      "sizeBytes": 81920,
-      "lineCount": 1427,
-      "originalFormat": "json",
-      "appliedPolicies": ["huge-output", "summary"],
-      "note": "Raw tool output stored as a temporary tool artifact.",
-      "summary": "Short optional summary."
-    }
-  }
-}
-```
-
-This is distinct from durable `agent_artifact_*` references. ToolGateway
-artifact refs are only for temporary tool output management.
-
-### Design choices
-
-- Keep durable and temporary artifacts separate: session artifacts are part of
-  the product model, while tool artifacts are engine-owned scratch storage.
-- Store one canonical text payload: this keeps `get` and `search` simple and
-  avoids format-specific readers in the gateway.
-- Preserve the full raw output first: summarization is an additional view over
-  the same stored payload, not a destructive rewrite.
-- Prefer bounded follow-up tools over reinlining content: once a result is
-  artifactized, the parent agent should inspect only the parts it needs.
-
 ## Core Model
 
 The frontend artifact shape is defined in
