+++ src-tauri/src/todos.rs
@@ -12,7 +12,6 @@ use uuid::Uuid;
 
 const TODO_STATES: [&str; 4] = ["todo", "doing", "blocked", "done"];
 const TODO_NOTE_KINDS: [&str; 2] = ["learned", "critical"];
-const TODO_NOTE_SOURCES: [&str; 2] = ["agent", "context_gateway"];
 const MUTATION_INTENTS: [&str; 10] = [
     "exploration",
     "implementation",
@@ -138,32 +137,6 @@ pub struct TodoRecordMutationInput {
     pub paths: Vec<String>,
 }
 
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct TodoContextEnrichmentTodoUpdateInput {
-    pub todo_id: String,
-    #[serde(default)]
-    pub verify_things_learned_note_ids: Vec<String>,
-    #[serde(default)]
-    pub verify_critical_info_note_ids: Vec<String>,
-    #[serde(default)]
-    pub append_things_learned: Vec<String>,
-    #[serde(default)]
-    pub append_critical_info: Vec<String>,
-    #[serde(default)]
-    pub remove_duplicate_things_learned_note_ids: Vec<String>,
-    #[serde(default)]
-    pub remove_duplicate_critical_info_note_ids: Vec<String>,
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct TodoContextEnrichmentInput {
-    pub turn: i64,
-    #[serde(default)]
-    pub updates: Vec<TodoContextEnrichmentTodoUpdateInput>,
-}
-
 fn todo_store_root() -> Result<PathBuf, String> {
     Ok(app_store_root()?.join("sessions").join("todos"))
 }
@@ -248,108 +221,6 @@ fn normalize_paths(paths: &[String]) -> Result<Vec<String>, String> {
         }
     }
     Ok(out)
-}
-
-fn normalize_note_match_key(text: &str) -> String {
-    let collapsed = text.trim().split_whitespace().collect::<Vec<_>>().join(" ");
-    collapsed
-        .to_lowercase()
-        .trim_end_matches(|c| matches!(c, '.' | '!' | '?'))
-        .to_string()
-}
-
-fn notes_are_duplicates(left: &str, right: &str) -> bool {
-    left == right || normalize_note_match_key(left) == normalize_note_match_key(right)
-}
-
-fn dedupe_string_ids(ids: &[String], field_name: &str) -> Result<Vec<String>, String> {
-    let mut seen = HashSet::new();
-    let mut out = Vec::new();
-    for id in ids {
-        let normalized = normalize_required_text(id, field_name)?;
-        if seen.insert(normalized.clone()) {
-            out.push(normalized);
-        }
-    }
-    Ok(out)
-}
-
-fn apply_note_enrichment(
-    notes: &mut Vec<TodoNoteItem>,
-    verify_ids: &[String],
-    append_texts: &[String],
-    remove_duplicate_ids: &[String],
-    note_bucket: &str,
-    turn: i64,
-) -> Result<bool, String> {
-    let verify_ids = dedupe_string_ids(verify_ids, &format!("{}.verifyIds", note_bucket))?;
-    let remove_duplicate_ids = dedupe_string_ids(
-        remove_duplicate_ids,
-        &format!("{}.removeDuplicateIds", note_bucket),
-    )?;
-
-    let mut changed = false;
-
-    for note_id in &verify_ids {
-        let index = notes
-            .iter()
-            .position(|note| note.id == *note_id)
-            .ok_or_else(|| format!("NOT_FOUND: Note {} not found in {}", note_id, note_bucket))?;
-        if !notes[index].verified {
-            notes[index].verified = true;
-            changed = true;
-        }
-    }
-
-    let mut removal_indexes = Vec::new();
-    for note_id in &remove_duplicate_ids {
-        let index = notes
-            .iter()
-            .position(|note| note.id == *note_id)
-            .ok_or_else(|| format!("NOT_FOUND: Note {} not found in {}", note_id, note_bucket))?;
-        let note_text = notes[index].text.clone();
-        let has_earlier_duplicate = notes
-            .iter()
-            .take(index)
-            .any(|existing| notes_are_duplicates(existing.text.as_str(), note_text.as_str()));
-        if !has_earlier_duplicate {
-            return Err(format!(
-                "INVALID_ARGUMENT: Note {} in {} is not a later exact/near-exact duplicate",
-                note_id, note_bucket
-            ));
-        }
-        removal_indexes.push(index);
-    }
-
-    removal_indexes.sort_unstable();
-    removal_indexes.dedup();
-    for index in removal_indexes.into_iter().rev() {
-        notes.remove(index);
-        changed = true;
-    }
-
-    for text in append_texts {
-        let normalized_text =
-            normalize_required_text(text, &format!("{}.appendText", note_bucket))?;
-        if notes
-            .iter()
-            .any(|existing| notes_are_duplicates(existing.text.as_str(), normalized_text.as_str()))
-        {
-            continue;
-        }
-        let source = validate_choice("context_gateway", &TODO_NOTE_SOURCES, "source")?;
-        notes.push(TodoNoteItem {
-            id: new_short_id(),
-            text: normalized_text,
-            added_turn: turn,
-            author: "context_gateway".to_string(),
-            source,
-            verified: true,
-        });
-        changed = true;
-    }
-
-    Ok(changed)
 }
 
 fn load_todos_from_disk(path: &Path) -> Vec<TodoItem> {
@@ -776,97 +647,6 @@ pub fn todo_store_record_mutation(
     result
 }
 
-#[tauri::command]
-pub fn todo_store_context_enrich(
-    session_id: String,
-    input: TodoContextEnrichmentInput,
-    app_handle: AppHandle,
-    state: State<'_, AppState>,
-) -> Result<TodoMutationResponse, String> {
-    let start = Instant::now();
-    tool_log(
-        "todo_store_context_enrich",
-        "start",
-        json!({ "sessionId": session_id, "updateCount": input.updates.len() }),
-    );
-    let result = mutate_todos(&session_id, &state, |items| {
-        let mut seen_todo_ids = HashSet::new();
-        let mut touched_ids = HashSet::new();
-
-        for update in &input.updates {
-            let todo_id = normalize_required_text(&update.todo_id, "todoId")?;
-            if !seen_todo_ids.insert(todo_id.clone()) {
-                return Err(format!(
-                    "INVALID_ARGUMENT: duplicate todoId {} in context enrichment",
-                    todo_id
-                ));
-            }
-
-            let index = items
-                .iter()
-                .position(|item| item.id == todo_id)
-                .ok_or_else(|| format!("NOT_FOUND: Todo {} not found", todo_id))?;
-            let item = items.get_mut(index).unwrap();
-
-            let learned_changed = apply_note_enrichment(
-                &mut item.things_learned,
-                &update.verify_things_learned_note_ids,
-                &update.append_things_learned,
-                &update.remove_duplicate_things_learned_note_ids,
-                "thingsLearned",
-                input.turn,
-            )?;
-            let critical_changed = apply_note_enrichment(
-                &mut item.critical_info,
-                &update.verify_critical_info_note_ids,
-                &update.append_critical_info,
-                &update.remove_duplicate_critical_info_note_ids,
-                "criticalInfo",
-                input.turn,
-            )?;
-
-            if learned_changed || critical_changed {
-                item.last_touched_turn = input.turn;
-                touched_ids.insert(item.id.clone());
-            }
-        }
-
-        Ok(build_response(items, None, !touched_ids.is_empty()))
-    });
-
-    match &result {
-        Ok(response) => {
-            if response.removed {
-                emit_todo_changed(
-                    &app_handle,
-                    &TodoChangeEvent {
-                        session_id: session_id.clone(),
-                        todo_id: None,
-                        change: "context_enriched".to_string(),
-                        changed_at: now_ms(),
-                    },
-                );
-            }
-            tool_log(
-                "todo_store_context_enrich",
-                "ok",
-                json!({
-                    "sessionId": session_id,
-                    "changed": response.removed,
-                    "durationMs": start.elapsed().as_millis() as u64
-                }),
-            );
-        }
-        Err(error) => tool_log(
-            "todo_store_context_enrich",
-            "err",
-            json!({ "sessionId": session_id, "error": error, "durationMs": start.elapsed().as_millis() as u64 }),
-        ),
-    }
-
-    result
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -885,17 +665,6 @@ mod tests {
             critical_info: Vec::new(),
             mutation_log: Vec::new(),
             completion_note: None,
-        }
-    }
-
-    fn make_note(id: &str, text: &str, verified: bool) -> TodoNoteItem {
-        TodoNoteItem {
-            id: id.to_string(),
-            text: text.to_string(),
-            added_turn: 1,
-            author: "main".to_string(),
-            source: "agent".to_string(),
-            verified,
         }
     }
 
@@ -969,57 +738,5 @@ mod tests {
                 "src/agent/runner.ts".to_string()
             ]
         );
-    }
-
-    #[test]
-    fn test_apply_note_enrichment_verifies_appends_and_dedupes_near_exact_duplicates() {
-        let mut notes = vec![
-            make_note("note-1", "Preserve the leading system prompt.", false),
-            make_note("note-2", "  preserve   the leading system prompt! ", false),
-        ];
-
-        let changed = apply_note_enrichment(
-            &mut notes,
-            &vec!["note-1".to_string()],
-            &vec![
-                "The compactor replaces apiMessages but keeps chatMessages.".to_string(),
-                "the compactor replaces apiMessages but keeps chatMessages!".to_string(),
-            ],
-            &vec!["note-2".to_string()],
-            "thingsLearned",
-            7,
-        )
-        .unwrap();
-
-        assert!(changed);
-        assert_eq!(notes.len(), 2);
-        assert!(notes[0].verified);
-        assert_eq!(notes[1].source, "context_gateway".to_string());
-        assert!(notes[1].verified);
-        assert_eq!(
-            notes[1].text,
-            "The compactor replaces apiMessages but keeps chatMessages."
-        );
-    }
-
-    #[test]
-    fn test_apply_note_enrichment_rejects_non_duplicate_removal() {
-        let mut notes = vec![
-            make_note("note-1", "First fact", false),
-            make_note("note-2", "Second fact", false),
-        ];
-
-        let error = apply_note_enrichment(
-            &mut notes,
-            &Vec::new(),
-            &Vec::new(),
-            &vec!["note-2".to_string()],
-            "criticalInfo",
-            7,
-        )
-        .unwrap_err();
-
-        assert!(error.contains("not a later exact/near-exact duplicate"));
-        assert_eq!(notes.len(), 2);
     }
 }
