+++ src/agent/types.ts
@@ -97,7 +97,7 @@ export interface AgentPlan {
 
 export type TodoStatus = "todo" | "doing" | "done" | "blocked";
 export type TodoOwner = "main" | string;
-export type TodoNoteSource = "agent" | "context_gateway";
+export type TodoNoteSource = "agent";
 export type MutationIntent =
   | "exploration"
   | "implementation"
@@ -179,8 +179,6 @@ export interface ToolCallDisplay {
   id: string;
   tool: string;
   args: Record<string, unknown>;
-  /** Internal ToolGateway phase for UI-only progress hints. */
-  gatewayPhase?: "summarizing";
   mcp?: {
     serverId: string;
     serverName: string;
